// src/atis/infrastructure/service/weather/parser/realtimetxt/weather-realtime.dto.ts
export interface WeatherRealtimeDto {
  readonly date: string;
  readonly time: string;
  readonly temp: string;
  readonly hum: string;
  readonly dew: string;
  readonly wspeed: string;
  readonly wlatest: string;
  readonly bearing: string;
  readonly rrate: string;
  readonly rfall: string;
  readonly press: string;
  readonly currentwdir: string;
  readonly beaufortnumber: string;
  readonly windunit: string;
  readonly tempunitnodeg: string;
  readonly pressunit: string;
  readonly rainunit: string;
  readonly windrun: string;
  readonly presstrendval: string;
  readonly rmonth: string;
  readonly ryear: string;
  readonly rfallY: string;
  readonly intemp: string;
  readonly inhum: string;
  readonly wchill: string;
  readonly temptrend: string;
  readonly tempTH: string;
  readonly TtempTH: string;
  readonly tempTL: string;
  readonly TtempTL: string;
  readonly windTM: string;
  readonly TwindTM: string;
  readonly wgustTM: string;
  readonly TwgustTM: string;
  readonly pressTH: string;
  readonly TpressTH: string;
  readonly pressTL: string;
  readonly TpressTL: string;
  readonly version: string;
  readonly build: string;
  readonly wgust: string;
  readonly heatindex: string;
  readonly humidex: string;
  readonly UV: string;
  readonly ET: string;
  readonly SolarRad: string;
  readonly avgbearing: string;
  readonly rhour: string;
  readonly forecastnumber: string;
  readonly isdaylight: string;
  readonly SensorContactLost: string;
  readonly wdir: string;
  readonly cloudbasevalue: string;
  readonly cloudbaseunit: string;
  readonly apptemp: string;
  readonly SunshineHours: string;
  readonly CurrentSolarMax: string;
  readonly IsSunny: string;
}